Integrate new bridged version of kinect into kinect class and factories
