+++ perception/kinect2_sensor.py
@@ -6,11 +6,21 @@ import copy
 import logging
 import numpy as np
 import os
+import time
 
 try:
     import pylibfreenect2 as lf2
 except:
     logging.warning('Unable to import pylibfreenect2. Python-only Kinect driver may not work properly.')
+
+try:
+    from cv_bridge import CvBridge, CvBridgeError
+    import rospy
+    import sensor_msgs.msg
+    import sensor_msgs.point_cloud2 as pc2
+except ImportError:
+    logging.warning("Failed to import ROS in Kinect2_sensor.py. Kinect will not be able to be used in bridged mode")
+    
 
 from .constants import MM_TO_METERS, INTR_EXTENSION
 from .camera_intrinsics import CameraIntrinsics
@@ -41,9 +51,17 @@ class Kinect2DepthMode:
     METERS = 0
     MILLIMETERS = 1
 
+class Kinect2BridgedQuality:
+    """Kinect depth mode setting for bridged mode
+    """
+    HD = "hd"
+    QUARTER_HD = "qhd"
+    SD = "sd"
+
 class Kinect2Sensor(CameraSensor):
     # constants for image height and width (in case they're needed somewhere)
-    """Class for interacting with a Kinect v2 RGBD sensor.
+    """Class for interacting with a Kinect v2 RGBD sensor directly through protonect driver.
+    https://github.com/OpenKinect/libfreenect2
     """
 
     #Constants for image height and width (in case they're needed somewhere)
@@ -56,8 +74,8 @@ class Kinect2Sensor(CameraSensor):
                  registration_mode = Kinect2RegistrationMode.COLOR_TO_DEPTH,
                  depth_mode = Kinect2DepthMode.METERS,
                  device_num=0, frame=None):
-        """Initialize a Kinect v2 sensor with the given configuration.
-
+        """Initialize a Kinect v2 sensor directly to the protonect driver with the given configuration.
+        When kinect is connected to the protonect driver directly, the iai_kinect kinect_bridge cannot be run at the same time
         Parameters
         ----------
         packet_pipeline_mode : int
@@ -319,6 +337,198 @@ class Kinect2Sensor(CameraSensor):
                 IrImage(ir_arr.astype(np.uint16), self._ir_frame),
                 color_depth_map)
 
+class KinectSensorBridged(CameraSensor):
+    """Class for interacting with a Kinect v2 RGBD sensor through the kinect bridge
+    https://github.com/code-iai/iai_kinect2. This is preferrable for visualization and debug
+    because the kinect bridge will continuously publish image and point cloud info.
+    """
+    QUALITY = Kinect2BridgedQuality.HD
+    TOPIC_IMAGE_COLOR = '/kinect2/%s/image_color' %(QUALITY)
+    TOPIC_IMAGE_DEPTH = '/kinect2/%s/image_depth_rect' %(QUALITY)
+    TOPIC_INFO_CAMERA = '/kinect2/%s/camera_info' %(QUALITY)
+
+    def __init__(self, quality=Kinect2BridgedQuality.HD, frame='kinect2_rgb_optical_frame'):
+        """Initialize a Kinect v2 sensor which connects to the iai_kinect2 bridge
+        ----------
+        quality : :obj:`str`
+            The quality (HD, Quarter-HD, SD) of the image data that should be subscribed to
+        frame : :obj:`str`
+            The name of the frame of reference in which the sensor resides.
+            If None, this will be set to 'kinect2_rgb_optical_frame'
+       """
+        # set member vars
+        self._frame = frame
+        self._quality = quality
+        
+        self._initialized = False
+        self._format = None
+        self._camera_intr = None
+        self._cur_depth_im = None
+        self._running = False
+        self._bridge = CvBridge()
+        
+    def __del__(self):
+        """Automatically stop the sensor for safety.
+        """
+        if self.is_running:
+            self.stop()
+            
+    def _set_camera_properties(self, msg):
+        """ Set the camera intrinsics from an info msg. """
+        focal_x = msg.K[0]
+        focal_y = msg.K[4]
+        center_x = msg.K[2]
+        center_y = msg.K[5]
+        im_height = msg.height
+        im_width = msg.width
+        self._camera_intr = CameraIntrinsics(self._frame, focal_x, focal_y,
+                                             center_x, center_y,
+                                             height=im_height,
+                                             width=im_width)
+
+    def _process_image_msg(self, msg):
+        """ Process an image message and return a numpy array with the image data
+        Returns
+        -------
+        :obj:`numpy.ndarray` containing the image in the image message
+
+        Raises
+        ------
+        CvBridgeError
+            If the bridge is not able to convert the image
+        """
+        encoding = msg.encoding
+        try:
+            image = self._bridge.imgmsg_to_cv2(msg, encoding)
+        except CvBridgeError as e:
+            rospy.logerr(e)
+        return image
+        
+    def _color_image_callback(self, image_msg):
+        """ subscribe to image topic and keep it up to date
+        """
+        color_arr = self._process_image_msg(image_msg)
+        self._cur_color_im = ColorImage(color_arr[:,:,::-1], self._frame)
+ 
+    def _depth_image_callback(self, image_msg):
+        """ subscribe to depth image topic and keep it up to date
+        """
+        depth_arr = self._process_image_msg(image_msg)
+        depth = np.array(depth_arr, np.float32)
+        self._cur_depth_im = DepthImage(depth, self._frame)
+        
+    def _camera_info_callback(self, msg):
+        """ Callback for reading camera info. """
+        self._camera_info_sub.unregister()
+        self._set_camera_properties(msg)
+    
+    @property
+    def ir_intrinsics(self):
+        """:obj:`CameraIntrinsics` : The camera intrinsics for the Ensenso IR camera.
+        """
+        return self._camera_intr
+
+    @property
+    def is_running(self):
+        """bool : True if the stream is running, or false otherwise.
+        """
+        return self._running
+
+    @property
+    def frame(self):
+        """:obj:`str` : The reference frame of the sensor.
+        """
+        return self._frame
+
+    def start(self):
+        """ Start the sensor """
+        # initialize subscribers
+        self._image_sub = rospy.Subscriber(KinectSensorBridged.TOPIC_IMAGE_COLOR, sensor_msgs.msg.Image, self._color_image_callback)
+        self._depth_sub = rospy.Subscriber(KinectSensorBridged.TOPIC_IMAGE_DEPTH, sensor_msgs.msg.Image, self._depth_image_callback)
+        self._camera_info_sub = rospy.Subscriber(KinectSensorBridged.TOPIC_INFO_CAMERA, sensor_msgs.msg.CameraInfo, self._camera_info_callback)
+        
+        timeout = 10
+        try:
+            logging.info("waiting to recieve a message from the Kinect")
+            rospy.wait_for_message(KinectSensorBridged.TOPIC_IMAGE_COLOR, sensor_msgs.msg.Image, timeout=timeout)
+            rospy.wait_for_message(KinectSensorBridged.TOPIC_IMAGE_DEPTH, sensor_msgs.msg.Image, timeout=timeout)
+            rospy.wait_for_message(KinectSensorBridged.TOPIC_INFO_CAMERA, sensor_msgs.msg.CameraInfo, timeout=timeout)
+        except rospy.ROSException as e:
+            logging.error("Kinect topic not found, Kinect not started")
+            logging.error(e)
+
+        while self._camera_intr is None:
+            time.sleep(0.1)
+        
+        self._running = True
+
+    def stop(self):
+        """ Stop the sensor """
+        # check that everything is running
+        if not self._running:
+            logging.warning('Kinect not running. Aborting stop')
+            return False
+
+        # stop subs
+        self._image_sub.unregister()
+        self._depth_sub.unregister()
+        self._camera_info_sub.unregister
+
+        self._running = False
+        return True
+
+    def frames(self):
+        """Retrieve a new frame from the Ensenso and convert it to a ColorImage,
+        a DepthImage, IrImage is always none for this type
+
+        Returns
+        -------
+        :obj:`tuple` of :obj:`ColorImage`, :obj:`DepthImage`, :obj:`IrImage`, :obj:`numpy.ndarray`
+            The ColorImage, DepthImage, and IrImage of the current frame.
+
+        Raises
+        ------
+        RuntimeError
+            If the Kinect stream is not running.
+        """
+        # wait for a new image
+        while self._cur_depth_im is None or self._cur_color_im is None:
+            time.sleep(0.01)
+            
+        # read next image
+        depth_im = self._cur_depth_im
+        color_im = self._cur_color_im
+
+        self._cur_color_im = None
+        self._cur_depth_im = None
+
+        #TODO add ir image
+        return color_im, depth_im, None
+
+    def median_depth_img(self, num_img=1, fill_depth=0.0):
+        """Collect a series of depth images and return the median of the set.
+
+        Parameters
+        ----------
+        num_img : int
+            The number of consecutive frames to process.
+
+        Returns
+        -------
+        :obj:`DepthImage`
+            The median DepthImage collected from the frames.
+        """
+        depths = []
+
+        for _ in range(num_img):
+            _, depth, _ = self.frames()
+            depths.append(depth)
+
+        median_depth = Image.median_images(depths)
+        median_depth.data[median_depth.data == 0.0] = fill_depth
+        return median_depth
+  
+
 class VirtualKinect2Sensor(CameraSensor):
     """Class for a virtual Kinect v2 sensor that uses pre-captured images
     stored to disk instead of actually connecting to a sensor.
@@ -534,6 +744,8 @@ class Kinect2SensorFactory:
         elif sensor_type == 'virtual':
             s = VirtualKinect2Sensor(cfg['image_dir'],
                                      frame=cfg['frame'])
+        elif sensor_type == 'bridged':
+            s = KinectSensorBridged(quality=cfg['quality'], frame=cfg['frame'])
         else:
             raise ValueError('Kinect2 sensor type %s not supported' %(sensor_type))
         return s
